refactor: extract FloatingToc component from CodemirrorEditor
